rename search -> selection
